pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img2pose"
version = "1.0.0"
description = "Real-time 6DoF face pose estimation - face detection and alignment without prior face detection"
readme = "src/img2pose/README.md"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Vitor Albiero", email = "vitor.albiero@gmail.com"},
]
maintainers = [
    {name = "Vitor Albiero", email = "vitor.albiero@gmail.com"},
]
keywords = [
    "face-detection",
    "face-alignment",
    "pose-estimation",
    "6dof",
    "deep-learning",
    "computer-vision",
    "pytorch",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=1.7.0",
    "torchvision>=0.8.0",
    "numpy>=1.19.0",
    "Pillow>=8.0.0",
    "scipy>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]
visualization = [
    "opencv-python>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/vitoralbiero/img2pose"
Documentation = "https://github.com/vitoralbiero/img2pose#readme"
Repository = "https://github.com/vitoralbiero/img2pose"
Issues = "https://github.com/vitoralbiero/img2pose/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
img2pose = ["data/*.npy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --import-mode=importlib"
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
